tools/requirements.txt
typer>=0.9.0
rich>=13.0.0
tomli>=2.0.0
Pillow>=10.0.0
